Middleware done?

+++ controllers/teas.py
@@ -1,12 +1,12 @@
 from fastapi import APIRouter, Depends, HTTPException
 from sqlalchemy.orm import Session
 from models.tea import TeaModel
-from serializers.tea import TeaSchema
 from models.user import UserModel
+from serializers.tea import TeaSchema, TeaCreate as TeaCreateSchema
+
 from typing import List
 from database import get_db
 from dependencies.get_current_user import get_current_user
-
 
 router = APIRouter()
 
@@ -15,7 +15,6 @@ def get_teas(db: Session = Depends(get_db)):
     teas = db.query(TeaModel).all()
     return teas
 
-
 @router.get("/teas/{tea_id}", response_model=TeaSchema)
 def get_single_tea(tea_id: int, db: Session = Depends(get_db)):
     tea = db.query(TeaModel).filter(TeaModel.id == tea_id).first()
@@ -23,21 +22,22 @@ def get_single_tea(tea_id: int, db: Session = Depends(get_db)):
         raise HTTPException(status_code=404, detail="Tea not found")
     return tea
 
-
 @router.post("/teas", response_model=TeaSchema)
-def create_tea(tea: TeaSchema, db: Session = Depends(get_db), current_user: UserModel = Depends(get_current_user)):
-    new_tea = TeaModel(**tea.dict())
+def create_tea(tea: TeaCreateSchema, db: Session = Depends(get_db), current_user: UserModel = Depends(get_current_user)):
+    new_tea = TeaModel(**tea.dict(), user_id=current_user.id)
     db.add(new_tea)
     db.commit()
     db.refresh(new_tea)
     return new_tea
 
-
 @router.put("/teas/{tea_id}", response_model=TeaSchema)
-def update_tea(tea_id: int, tea: TeaSchema, db: Session = Depends(get_db), current_user: UserModel = Depends(get_current_user)):  # Add the current_user dependency
+def update_tea(tea_id: int, tea: TeaSchema, db: Session = Depends(get_db), current_user: UserModel = Depends(get_current_user)):
     db_tea = db.query(TeaModel).filter(TeaModel.id == tea_id).first()
     if not db_tea:
         raise HTTPException(status_code=404, detail="Tea not found")
+
+    if db_tea.user_id != current_user.id:
+        raise HTTPException(status_code=403, detail="Operation forbidden")
 
     tea_data = tea.dict(exclude_unset=True)
     for key, value in tea_data.items():
@@ -47,12 +47,14 @@ def update_tea(tea_id: int, tea: TeaSchema, db: Session = Depends(get_db), curre
     db.refresh(db_tea)
     return db_tea
 
-
 @router.delete("/teas/{tea_id}")
-def delete_tea(tea_id: int, db: Session = Depends(get_db), current_user: UserModel = Depends(get_current_user)):  # Add the current_user dependency
+def delete_tea(tea_id: int, db: Session = Depends(get_db), current_user: UserModel = Depends(get_current_user)):
     db_tea = db.query(TeaModel).filter(TeaModel.id == tea_id).first()
     if not db_tea:
         raise HTTPException(status_code=404, detail="Tea not found")
+
+    if db_tea.user_id != current_user.id:
+        raise HTTPException(status_code=403, detail="Operation forbidden")
 
     db.delete(db_tea)
     db.commit()
